backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-auth-oauthlib
google-auth-transport-requests
oauthlib
python-jose
passlib[bcrypt]
google-generativeai
anyio
starlette
